+++ migrations/004_rollback_clientes_table.sql
@@ -1,0 +1,9 @@
+-- Rollback: Drop clientes table
+-- Description: Remove tabela de clientes
+-- Date: 2025-11-04
+
+BEGIN;
+
+DROP TABLE IF EXISTS clientes CASCADE;
+
+COMMIT;